requirements.txt
streamlit>=1.28.0
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
pandas>=1.5.0
numpy<2.0,>=1.24.0 # numpy>=1.24.0
Pillow>=9.5.0
pypdf2>=3.0.0
pdfplumber>=0.9.0
python-multipart>=0.0.6
pdf2image>=1.17.0
pdfminer.six>=20221105
tqdm>=4.65.0
rich>=13.0.0
langchain>=0.1.0
sentence-transformers>=2.2.2
langchain-community
faiss-cpu
scikit-learn>=1.3.0 